今日はカレーをたべましたこみっとっっ！！
