+++ launcher.py
@@ -36,22 +36,23 @@ async def playing(ctx, title):
 init()
 
 bot.load_extension("dispander")  # diapanderをextensionとして読み込み
-# bot.load_extension("cogs.tanaka")
+bot.load_extension("cogs.apex_tracker")
+bot.load_extension("cogs.commandslist")
+bot.load_extension("cogs.spotify")
+bot.load_extension("cogs.vcwhite")
 bot.load_extension("cogs.marimo")
+bot.load_extension("cogs.what_today")
+bot.load_extension("cogs.save_image")
+# bot.load_extension("cogs.tanaka")
 # bot.load_extension("cogs.read")
 # bot.load_extension("cogs.policezen")
 # bot.load_extension("cogs.policehan")
 # bot.load_extension("cogs.react")
 # bot.load_extension("cogs.yobro")
-# bot.load_extension("cogs.vcalert")
-bot.load_extension("cogs.vctest")
+# bot.load_extension("cogs.vctest")
 # bot.load_extension("cogs.greet")
 # bot.load_extension("cogs.notify")
-bot.load_extension("cogs.what_today")
-bot.load_extension("cogs.save_image")
 # bot.load_extension("cogs.delete_image")
-bot.load_extension("cogs.apex_tracker")
-bot.load_extension("cogs.commandslist")
-bot.load_extension("cogs.spotify")
+
 
 bot.run(TOKEN)
